🩹 :: move preferenceManager data to presentation
