+++ presentation/src/main/java/kr/hs/dgsw/bamboo/bamboo_android/util/BambooApplication.kt
@@ -2,7 +2,6 @@ package kr.hs.dgsw.bamboo.bamboo_android.util
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
-import kr.hs.dgsw.bamboo.data.database.prefs.PreferenceManager
 
 @HiltAndroidApp
 class BambooApplication : Application() {
